requirements.txt
streamlit
paho-mqtt
pandas
plotly
numpy
